Interface Example
New field is added in item card called 'Item Classification Type'. This is an extensible enum that holds values Class A and Class B. If a user selects 'Item Classification Type' - Class A, the interface inserts a new comment in comment line page. If Class B is selected, a message pops us saying ' classification code of the item is Class B'.